Started to comment the code, some class reorganization, a few security fixes.
