+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/RouteCostPrinterColumn.java
@@ -3,7 +3,7 @@ package com.graphhopper.jsprit.core.reporting.route;
 import java.util.function.Consumer;
 
 import com.graphhopper.jsprit.core.problem.solution.route.activity.Start;
-import com.graphhopper.jsprit.core.reporting.DynamicTableDefinition.ColumnDefinition.Builder;
+import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnDefinition;
 
 public class RouteCostPrinterColumn extends TransportCostPrinterColumn {
 
@@ -13,7 +13,7 @@ public class RouteCostPrinterColumn extends TransportCostPrinterColumn {
         super();
     }
 
-    public RouteCostPrinterColumn(Consumer<Builder> decorator) {
+    public RouteCostPrinterColumn(Consumer<ColumnDefinition.Builder> decorator) {
         super(decorator);
     }
