+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/util/VehicleIndexComparator.java
@@ -18,9 +18,9 @@
 
 package com.graphhopper.jsprit.core.util;
 
-import com.graphhopper.jsprit.core.problem.solution.route.VehicleRoute;
-
 import java.util.Comparator;
+
+import com.graphhopper.jsprit.core.problem.solution.route.VehicleRoute;
 
 public class VehicleIndexComparator implements Comparator<VehicleRoute> {
     public int compare(VehicleRoute a, VehicleRoute b) {
